+++ src/hardis-plugins-provider.ts
@@ -16,7 +16,7 @@ let gitInstallOk = false;
 export class HardisPluginsProvider
   implements vscode.TreeDataProvider<StatusTreeItem>
 {
-  constructor(private workspaceRoot: string) {}
+  constructor(private workspaceRoot: string) { }
 
   getTreeItem(element: StatusTreeItem): vscode.TreeItem {
     return element;
@@ -48,10 +48,10 @@ export class HardisPluginsProvider
       topic.id === "status-plugins-sfdx"
         ? await this.getPluginsItems()
         : topic.id === "status-plugins-core"
-        ? await this.getCoreItems()
-        : topic.id === "status-vscode-extensions"
-        ? await this.getExtensionsItems()
-        : [];
+          ? await this.getCoreItems()
+          : topic.id === "status-vscode-extensions"
+            ? await this.getExtensionsItems()
+            : [];
     console.timeEnd("TreeViewItem_init_" + topic.id);
     Logger.log("Completed TreeViewItem_init_" + topic.id);
     for (const item of topicItems) {
@@ -211,6 +211,10 @@ export class HardisPluginsProvider
     // Check sfdx related installs
     const plugins = [
       {
+        name: "@salesforce/plugin-packaging",
+        helpUrl: "https://www.npmjs.com/package/@salesforce/plugin-packaging",
+      },
+      {
         name: "sfdx-hardis",
         helpUrl: "https://hardisgroupcom.github.io/sfdx-hardis/",
       },
@@ -300,7 +304,7 @@ export class HardisPluginsProvider
         sfdxCliOutdated = true;
         sfdxCliItem.label =
           sfdxCliItem.label.includes("missing") &&
-          !sfdxCliItem.label.includes("(link)")
+            !sfdxCliItem.label.includes("(link)")
             ? sfdxCliItem.label
             : sfdxCliItem.label + " (upgrade available)";
         sfdxCliItem.command = `npm install @salesforce/cli@${recommendedSfdxCliVersion} -g`;
@@ -344,7 +348,7 @@ export class HardisPluginsProvider
       if (!sfdxPlugins.includes(`${plugin.name} ${latestPluginVersion}`)) {
         pluginItem.label =
           pluginItem.label.includes("missing") &&
-          !pluginItem.label.includes("(link)")
+            !pluginItem.label.includes("(link)")
             ? pluginItem.label.replace("(link)", "(localdev)")
             : pluginItem.label + " (upgrade available)";
         pluginItem.command = `echo y|sfdx plugins:install ${plugin.name} && sfdx hardis:work:ws --event refreshPlugins`;